# application/docs/swagger.yaml
basePath: /
definitions:
  requests.CreateUserRequest:
    properties:
      e-mail:
        example: andrew123@qwerty.kom
        type: string
      login:
        example: TopPlayer123
        type: string
      nickname:
        example: Looser1123
        type: string
      password:
        example: 123superPassword
        type: string
    required:
    - e-mail
    - login
    - nickname
    - password
    type: object
  requests.CreateUserResponse:
    properties:
      id:
        example: 2
        type: integer
      session:
        $ref: '#/definitions/requests.RefreshSessionResponse'
    type: object
  requests.RefreshSessionRequest:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  requests.RefreshSessionResponse:
    properties:
      accessToken:
        type: string
      expiresAt:
        type: integer
      refreshToken:
        type: string
    type: object
  requests.SignInRequest:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
  requests.SignInResponse:
    properties:
      accessToken:
        type: string
      expiresAt:
        type: integer
      refreshToken:
        type: string
    type: object
  requests.VerificationRequest:
    properties:
      code:
        type: string
    required:
    - code
    type: object
host: localhost:8080
info:
  contact: {}
  description: I hate this app.
  title: Smartri Backend
  version: 1.3.3.7
paths:
  /auth/signin:
    post:
      consumes:
      - application/json
      description: sign in
      parameters:
      - description: request format
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.SignInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/requests.SignInResponse'
      summary: sign in
  /auth/signup:
    post:
      consumes:
      - application/json
      description: регистрация нового пользователя
      parameters:
      - description: request format
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/requests.CreateUserResponse'
      summary: регистрация нового пользователя
  /auth/token/update:
    post:
      consumes:
      - application/json
      description: refreshes current session with token pair (access + refresh)
      parameters:
      - description: request format
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.RefreshSessionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/requests.RefreshSessionResponse'
      summary: refreshes session
  /user/verify:
    post:
      consumes:
      - application/json
      description: AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAaa
      parameters:
      - description: request format
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.VerificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Ok
      summary: verifies user
swagger: "2.0"